change Interpolate to Preprocess
This function is now doing more than just an interpolation. Also,
special characters in strings are ignored. Therefore, it is possible
to execute queries like:
	* SELECT 'non[]sence'
	* SELECT 'nonsence?'
without errors or char. replacing. ([] would be otherwise converted
to `` and ? would be interpolated.)

NOTE: It is not an ideal implementation yet. There is no escaping of
characters in strings written like this, so 'dummy\'' won't work.