improve methods Where and Having
This CL makes a huge change of behaviour of Where (and Having) method.
Now, only possible types of the first argument are string or And type.
Type And replaces original Eq type. It is because And goes beyond an
equality map. When an And map is the input, it is only a shorthand for
calling the Where (Having) method multiple times with 2 arguments:
an expression string and a parameter.

Example:
	b.Where(ql.And{"name": "Igor", "age >": 50})
	// is equivalent to:
	b.Where("name", "Igor")
	b.Where("age >", 50)

It introduces the "short notation":
	column_name [operator] [?]
Default operator is '=', '?' placeholder is not required. If the
expression doesn't match this format it is processed as usual.

There might be some TODOs with defining the right regexp for the short
notation.